# api/openapi.yaml
openapi: 3.1.0
info:
  title: CiftPay API
  version: 1.0.0
  description: |
    CiftPay turns M-Pesa payments into KRA eTIMS invoices.

    Conventions:
    - All money values are integer **cents** (`amount_cents`). KES 2,400.00 = 240000.
    - All timestamps are RFC 3339 UTC.
    - Authenticated endpoints use the `ciftpay_session` HttpOnly cookie plus `X-CSRF-Token` on non-GET.
    - The active organisation is the caller's default membership or the `X-Org-Id` header.
    - Errors use the envelope in `#/components/schemas/Error`.
    - Personal data (MSISDN, KRA PIN) is returned masked unless stated otherwise.
  contact:
    name: CiftPay Engineering
servers:
  - url: http://localhost:8080/v1
    description: Local development
  - url: https://api.ciftpay.co.ke/v1
    description: Production
...
security:
  - sessionCookie: []
paths:
...
  /kra/pin/{pin}:
    parameters:
      - name: pin
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/KraPin"
    get:
      tags: [kra]
      operationId: checkKraPin
      security: []
      summary: Validate KRA PIN and resolve registered legal name
      responses:
        "200":
          description: Taxpayer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KraTaxpayer"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /kra/obligations/{pin}:
    parameters:
      - name: pin
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/KraPin"
    get:
      tags: [kra]
      operationId: fetchKraObligations
      security: []
      summary: Query active tax obligations for a KRA PIN
      responses:
        "200":
          description: List of obligations
          content:
            application/json:
              schema:
                type: object
                required: [pin, obligations]
                properties:
                  pin:
                    $ref: "#/components/schemas/KraPin"
                  obligations:
                    type: array
                    items:
                      $ref: "#/components/schemas/KraObligation"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ciftpay_session
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code, snake_case
              examples: [validation_failed, not_found, shortcode_claimed, rate_limited, unauthenticated, forbidden, conflict, internal]
            message:
              type: string
              description: Human-readable, safe to show to users
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
...
    KraPin:
      type: string
      pattern: "^[AP][0-9]{9}[A-Z]$"
      example: "A123456789B"
    KraTaxpayer:
      type: object
      required: [pin, taxpayer_name, taxpayer_status, vat_registered]
      properties:
        pin:
          $ref: "#/components/schemas/KraPin"
        taxpayer_name:
          type: string
          example: "SAFARICOM COMMERCIAL PARTNER LTD"
        taxpayer_status:
          type: string
          example: "ACTIVE"
        vat_registered:
          type: boolean
          example: true
        email:
          type: string
          nullable: true
    KraObligation:
      type: object
      required: [obligation_id, obligation_name, status]
      properties:
        obligation_id:
          type: string
          example: "VAT01"
        obligation_name:
          type: string
          example: "Value Added Tax (VAT)"
        effective_from:
          type: string
          example: "2021-06-01"
        status:
          type: string
          example: "ACTIVE"
...